requirements.txt
pyOpenSSL==16.0.0
boto==2.39.0
PyYAML==3.11
pycrypto==2.6.1
-e git+https://github.com/kuba/letsencrypt.git#egg=acme&subdirectory=acme
